Deprecate `.overrideClasspath()`
This is deprecated because the classpath may differ from the classpath
of the classloaders, which can lead to unpredictability (classes being
loaded twice into different classloaders, classes not being able to be
cast to their superclasses, classes not being found, etc.). Scanning
should always be driven by classloaders (which can provide classpath
elements), not by the classpath itself (which we can only rely on the
context classloader to handle, and that only works reliably if the
specified classpath is exactly the same as the context classloader's
classpath, which defeats the purpose of being able to specify an
override classpath).