Make Kotlin tests pass (#175, #183)
This is at the expense of displaying parameter names and parameter
modifiers, if their arity is different than the arity of the type
signature; it is also at the expense of displaying synthetic and/or
mandated parameters in many cases, since the type signature is used in
preference to the type descriptor (#183).

This is unavoidable if type parameter information is wanted, since type
parameter information is in the type signature but not the type
descriptor, but the type signature cannot always be aligned with the
type descriptor.