Implement loading ClassLoaderHandler from Java’s ServiceLoader.
This commit adds loading of additional ClassLoaderHandler from Java’s Serviceloader SPI.

The default list of handlers is now created in a static initializer block and than supplemented by a list of dynamically loaded handlers.
To use the the handlers from the SPI, ClassLoaderHandlerRegistryEntry had to have a new constructor, taking, taking in the pre populated handler.