Fix QuarkusClassLoaderHandler on Quarkus 2.7.
In Quarkus 2.7, the class loader stopped returning `JarClassPathElement`
or `DirectoryClassPathElement` entries. Instead, some implementation of
`ClassPathElement` is returned. This change tries to access it's
`getRoot` method, returning a `java.nio.file.Path` pointing to the root
of the element, as supported with class graph.

The behaviour has been verified on Quarkus 2.6 and Quarkus 2.7.